fix: float and Decimal error

+++ market_data/models.py
@@ -82,8 +82,8 @@ class StockPrice(models.Model):
         
         if latest_record:
             if latest_record.currency == "USD":
-                return latest_record.close_price * 1400
-            return latest_record.close_price
+                return float(latest_record.close_price * 1400)
+            return float(latest_record.close_price)
         
         # Step 2: 데이터가 없을 경우 fetch_and_save_stock_data 호출
         stock = yf.Ticker(symbol)
@@ -94,8 +94,8 @@ class StockPrice(models.Model):
         
         if latest_record:
             if latest_record.currency == "USD":
-                return latest_record.close_price * 1400
-            return latest_record.close_price
+                return float(latest_record.close_price * 1400)
+            return float(latest_record.close_price)
         
         # Step 4: fetch_and_save_stock_data 호출 후에도 데이터가 없는 경우 예외 처리
         raise ValueError(f"No stock data available for symbol {symbol} before {datetime}.")
